Added skeleton Companies page
This page is designed to show off the companies and employees that are using
Ruby in the North East. We should also have a page (or section) for freelancers
and contractors.

It can be fleshed out with logos and product information. In time it would be a
good idea to have member profiles that are linked through.